formatter/requirements.txt
pika==1.3.2
Spire.Doc==12.4.0
plum-dispatch==1.7.4